code out of util to reduce size for non-exp cases
